+++ src/idtrackerai/session.py
@@ -85,7 +85,7 @@ class Session:
     """List of paths to the different files the video is composed of.
     If the video is a single file, the list will have length 1"""
     number_of_animals: int = 0
-    intensity_ths: None | Sequence[int] = None
+    intensity_ths: None | Sequence[float] = None
     area_ths: None | Sequence[float] = None
     # bkg_model: None | np.ndarray = None
     name: str = ""